+++ locales/en/common/meta.js
@@ -2,7 +2,7 @@ var meta =
     {
     "lang"        : "en",
     "charset"     : "utf8",
-    "title"       : "Maple Flag Missions - We Make Good Games Great!",
+    "abstract"    : "Maple Flag Missions - We Make Good Games Great!",
     "description" : "We provide flight sim enthusiasts with the most realistic, exciting and entertaining combat flight simulator missions available.",
     "author"      : "Mattia Affabris"
     }