+++ public/javascripts/hardware.js
@@ -1,0 +1,7 @@
+var Hardware = function()
+  {
+  fadeParagraphs()
+  }
+
+var header = new Header("hardware")
+var hardware = new Hardware()