+++ routes/about.js
@@ -5,18 +5,18 @@ var config = require(rootDir + "bin/config")
 var meta = require(rootDir + "locales/" + config.lang + "/common/meta")
 var header = require(rootDir + "locales/" + config.lang + "/common/header")
 var footer = require(rootDir + "locales/" + config.lang + "/common/footer")
-var home = require(rootDir + "locales/" + config.lang + "/home")
-var content = merge(meta, header, footer, home)
+var about = require(rootDir + "locales/" + config.lang + "/about")
+var content = merge(meta, header, footer, about)
 
 // -----------------------------------------
 
 var router = express.Router()
 
-router.get("/home", function(req, res, next)
+router.get("/about", function(req, res, next)
     {
-    config.debug.routing("GET request -> /home")
+    config.debug.routing("GET request -> /about")
 
-    res.render("home", content)
+    res.render("about", content)
     })
 
 // -----------------------------------------