+++ routes/faq.js
@@ -2,9 +2,9 @@ var express = require("express")
 var merge = require("object-merge")
 var config = require(rootDir + "bin/config")
 
-var meta = require(rootDir + "locales/" + config.lang + "/meta")
-var header = require(rootDir + "locales/" + config.lang + "/header")
-var footer = require(rootDir + "locales/" + config.lang + "/footer")
+var meta = require(rootDir + "locales/" + config.lang + "/common/meta")
+var header = require(rootDir + "locales/" + config.lang + "/common/header")
+var footer = require(rootDir + "locales/" + config.lang + "/common/footer")
 var faq = require(rootDir + "locales/" + config.lang + "/faq")
 var content = merge(meta, header, footer, faq)